recent input updated

+++ src/redux/studentSlice.jsx
@@ -68,9 +68,10 @@ export const fetchStudent = createAsyncThunk(
    
   export const editExtraMark = createAsyncThunk(
     "editExtraMark",
-    async ({ id, mark,description }, { rejectWithValue }) => {  
+    async ({ id, mark,description,userId }, { rejectWithValue }) => {  
+      console.log(userId)
       try {
-        const response = await api.put(`/student/score/${id}`, {mark:mark,description:description});
+        const response = await api.put(`/student/score/${id}`, {mark:mark,description:description,userId:userId});
         return response.data;
       } catch (error) {
         return rejectWithValue(error.response.data);
@@ -80,10 +81,10 @@ export const fetchStudent = createAsyncThunk(
    
   export const deleteExtraMark = createAsyncThunk(
     "deleteExtraMark",
-    async ({ id }, { rejectWithValue }) => {
+    async ({ id,userId }, { rejectWithValue }) => {
     
       try {
-        const response = await api.delete(`/student/score/${id}`);
+        const response = await api.delete(`/student/score/${id}`,{data:{userId}});
         return response.data;
       } catch (error) {
         return rejectWithValue(error.response.data);
